requirements.txt
Django
django-autoslug
django-cors-headers
django-filter
django-rest-swagger
django-smart-selects
djangorestframework
djangorestframework-simplejwt
drf-yasg
gunicorn
Markdown
Pillow

